// api/libs/localize/src/application/types.ts
export type ProductTranslation = {
  description?: string
  slug?: string
}

export type SeoMetadataTranslation = {
  title?: string
  description?: string
  keywords?: string
  ogTitle?: string
  ogDescription?: string
}

export type CollectionTranslation = {
  name?: string
  slug?: string
  description?: string
}